refactor exception-handling and add PATCH endpoint

+++ src/main/java/se/iths/rest/StudentRest.java
@@ -2,6 +2,7 @@ package se.iths.rest;
 
 import se.iths.entity.Student;
 import se.iths.exception.ExceptionMessage;
+import se.iths.exception.SaveStudentException;
 import se.iths.service.StudentService;
 import se.iths.utils.JsonConverter;
 
@@ -9,7 +10,7 @@ import javax.inject.Inject;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import java.net.http.HttpResponse;
+import java.nio.charset.StandardCharsets;
 import java.util.List;
 
 @Path("students")
@@ -23,6 +24,14 @@ public class StudentRest {
     @Path("")
     @POST
     public Response createStudent(Student student) {
+        if (student.getFirstName() == null || student.getLastName() == null || student.getEmail() == null) {
+            throw new SaveStudentException(Response.status(Response.Status.BAD_REQUEST)
+//                JsonConverter-lösningen behövs ej, görs automatiskt av ".entity(new ExceptionMessage(400, "BAD_REQUEST", "Required data is missing. Student not created")"
+//                String message = JsonConverter.convertExceptionToJson(new ExceptionMessage(400, "BAD_REQUEST", "Required data is missing. Student not created"));
+                    .entity(new ExceptionMessage(400, "BAD_REQUEST", "First name, last name or email is missing. Student not created"))
+                    .type(MediaType.APPLICATION_JSON)
+                    .build());
+        }
         Student studentResult = studentService.createStudent(student);
         return Response.ok(studentResult).build();
     }
@@ -33,8 +42,8 @@ public class StudentRest {
         List<Student> students = studentService.getAllStudents();
         if (students.isEmpty()) {
             throw new WebApplicationException(Response.status(Response.Status.NOT_FOUND)
-                    .entity(JsonConverter.convertStringToJson("No students in database"))
-                    .type(MediaType.TEXT_PLAIN_TYPE)
+                    .entity(new ExceptionMessage(404, "NOT_FOUND", "No saved students"))
+                    .type(MediaType.APPLICATION_JSON)
                     .build());
         }
         return Response.ok(students).build();
@@ -46,8 +55,8 @@ public class StudentRest {
         Student foundStudent = studentService.getStudentById(id);
         if (foundStudent == null) {
             throw new WebApplicationException(Response.status(Response.Status.NOT_FOUND)
-                    .entity("Student with Id " + id + " was not found in database.")
-                    .type(MediaType.TEXT_PLAIN_TYPE).build());
+                    .entity(new ExceptionMessage(404, "NOT_FOUND", "Student with "  + id + " was not found"))
+                    .type(MediaType.APPLICATION_JSON).build());
         }
         return Response.ok(foundStudent).build();
     }
@@ -58,8 +67,8 @@ public class StudentRest {
         List<Student> foundStudents = studentService.getStudentsByLastName(lastName);
         if (foundStudents.isEmpty()) {
             throw new WebApplicationException(Response.status(Response.Status.NOT_FOUND)
-                    .entity("No students with last name " + lastName + " found in database.")
-                    .type(MediaType.TEXT_PLAIN_TYPE).build());
+                    .entity(new ExceptionMessage(404, "NOT_FOUND", "No students with last name " + lastName + " found."))
+                    .type(MediaType.APPLICATION_JSON).build());
         }
         return Response.ok(foundStudents).build();
     }
@@ -67,13 +76,41 @@ public class StudentRest {
     @Path("")
     @PUT
     public Response updateStudent(Student student) {
+        if (student.getId() == null || student.getFirstName() == null || student.getLastName() == null || student.getEmail() == null) {
+            throw new WebApplicationException(Response.status(Response.Status.BAD_REQUEST)
+                    .entity(new ExceptionMessage(400, "BAD_REQUEST", "Id is missing. Student not updated"))
+                    .type(MediaType.APPLICATION_JSON).build());
+        }
+        if (studentService.getStudentById(student.getId()) == null) {
+            throw new WebApplicationException(Response.status(Response.Status.NOT_FOUND)
+                    .entity(new ExceptionMessage(404, "NOT_FOUND", "Student with "  + student.getId() + " was not found"))
+                    .type(MediaType.APPLICATION_JSON).build());
+        }
         Student updatedStudent = studentService.updateStudent(student);
         return Response.ok(updatedStudent).build();
     }
 
     @Path("{id}")
+    @PATCH
+    public Response modifyStudent(@PathParam("id") Long id, Student student) {
+        if (getStudentById(id) == null) {
+            throw new WebApplicationException(Response.status(Response.Status.NOT_FOUND)
+                    .entity(new ExceptionMessage(404, "NOT_FOUND", "Student with "  + student.getId() + " was not found"))
+                    .type(MediaType.APPLICATION_JSON).build());
+        }
+        Student modifiedStudent = studentService.modifyStudent(id, student);
+        return Response.ok(modifiedStudent).build();
+    }
+
+    @Path("{id}")
     @DELETE
     public Response deleteStudent(@PathParam("id") Long id) {
+        Student studentToRemove = studentService.getStudentById(id);
+        if (studentToRemove == null) {
+            throw new WebApplicationException(Response.status(Response.Status.NOT_FOUND)
+                    .entity(new ExceptionMessage(404, "NOT_FOUND", "Student with ID " + id + " not found. No student removed"))
+                    .type(MediaType.APPLICATION_JSON).build());
+        }
         studentService.deleteStudent(id);
         return Response.ok().build();
     }
